Roll OrderItemClient class back to what it was before putting in IDisposable, and put IDisposable in BaseSendReceiveClient
