+++ PetStore.OrderItem.Client/Client/OrderItemClient.cs
@@ -2,7 +2,8 @@
 using PetStore.OrderItem.Client.Client.Interface;
 using PetStore.Shared.Helpers;
 using PetStore.Shared.QueMessages;
-using RabbitMQ.Client;
+using PetStore.Shared.RabbitMQ;
+using PetStore.Shared.RabbitMQ.Factorys;
 using RabbitMQ.Client.Events;
 using System;
 using System.Collections.Concurrent;
@@ -10,37 +11,21 @@ using System.Threading.Tasks;
 
 namespace PetStore.OrderItem.Client
 {
-    public class OrderItemClient : IDisposable, IOrderItemClient
+    public class OrderItemClient : BaseSendReceiveClient, IOrderItemClient
     {
-    //    https://gigi.nullneuron.net/gigilabs/tag/taskcompletionsource/
-
-
-       private  ConcurrentDictionary<string, TaskCompletionSource<OrderResponse>> _pendingMessages;
+        private readonly ConcurrentDictionary<string, TaskCompletionSource<OrderResponse>> _pendingMessages;
         private const string _requestQueueName = "OrderItem_RequestQueue";
         private const string _responseQueueName = "OrderItem_ResponseQueue";
         private const string _exchangeName = ""; // default exchange
-        private EventingBasicConsumer _consumer;
-        private IConnection _connection;
-        private IModel _channel;
-
-        private bool disposed = false;
 
         public OrderItemClient()
+            : base(RabbitMQConfigFactory.Create(), _requestQueueName, _responseQueueName, _exchangeName)
         {
-            var factory = new ConnectionFactory() { HostName = "localhost" };
-            _connection = factory.CreateConnection();
-            _channel = _connection.CreateModel();
-            _channel.QueueDeclare(_requestQueueName, true, false, false, null);
-            _channel.QueueDeclare(_responseQueueName, true, false, false, null);
-            _consumer = new EventingBasicConsumer(this._channel);
-            _consumer.Received += Receive;
-            _channel.BasicConsume(_responseQueueName, true, _consumer);
-            var replyQueueName = _channel.QueueDeclare().QueueName;
+            _pendingMessages = new ConcurrentDictionary<string, TaskCompletionSource<OrderResponse>>();
         }
 
         public Task<OrderResponse> Send(StockOrder stockOrder)
         {
-            _pendingMessages = new ConcurrentDictionary<string, TaskCompletionSource<OrderResponse>>();
             var message = stockOrder.Serialize();
             var tcs = new TaskCompletionSource<OrderResponse>();
             var correlationId = Guid.NewGuid().ToString();
@@ -49,46 +34,15 @@ namespace PetStore.OrderItem.Client
             return tcs.Task;
         }
 
-        private void Publish(byte[] message, string correlationId)
-        {
-            var props = this._channel.CreateBasicProperties();
-            props.CorrelationId = correlationId;
-            props.ReplyTo = _responseQueueName;
-            this._channel.BasicPublish(_exchangeName, _requestQueueName, props, message);
-        }
-
-        protected void Receive(object sender, BasicDeliverEventArgs e)
+        protected override void Receive(object sender, BasicDeliverEventArgs e)
         {
             var correlationId = e.BasicProperties.CorrelationId;
             var orderResponse = (OrderResponse)e.Body.ToArray().DeSerialize(typeof(OrderResponse));
-
-            //On the first time round _pendingMessages has one item in it, and is the item we are looking for
-            //On the second call it is empty
             this._pendingMessages.TryRemove(correlationId, out var tcs);
             if (tcs != null)
             {
                 tcs.SetResult(orderResponse);
             }
         }
-
-
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!disposed && disposing)
-            {
-                _channel?.Dispose();
-                _connection?.Dispose();
-            }
-
-            this.disposed = true;
-        }
     }
-
 }
